implement balancer policy
