+++ src/meta/greedy_load_balancer.cpp
@@ -35,6 +35,7 @@
 #include "greedy_load_balancer.h"
 #include "meta_data.h"
 #include "meta_admin_types.h"
+#include "app_balance_policy.h"
 
 namespace dsn {
 namespace replication {
@@ -102,131 +103,6 @@ void get_min_max_set(const std::map<rpc_address, uint32_t> &node_count_map,
     }
 }
 
-const std::string &get_disk_tag(const app_mapper &apps, const rpc_address &node, const gpid &pid)
-{
-    const config_context &cc = *get_config_context(apps, pid);
-    auto iter = cc.find_from_serving(node);
-    dassert(iter != cc.serving.end(),
-            "can't find disk tag of gpid(%d.%d) for %s",
-            pid.get_app_id(),
-            pid.get_partition_index(),
-            node.to_string());
-    return iter->disk_tag;
-}
-
-std::shared_ptr<configuration_balancer_request>
-generate_balancer_request(const app_mapper &apps,
-                          const partition_configuration &pc,
-                          const balance_type &type,
-                          const rpc_address &from,
-                          const rpc_address &to)
-{
-    FAIL_POINT_INJECT_F("generate_balancer_request", [](string_view name) { return nullptr; });
-
-    configuration_balancer_request result;
-    result.gpid = pc.pid;
-
-    std::string ans;
-    switch (type) {
-    case balance_type::MOVE_PRIMARY:
-        ans = "move_primary";
-        result.balance_type = balancer_request_type::move_primary;
-        result.action_list.emplace_back(
-            new_proposal_action(from, from, config_type::CT_DOWNGRADE_TO_SECONDARY));
-        result.action_list.emplace_back(
-            new_proposal_action(to, to, config_type::CT_UPGRADE_TO_PRIMARY));
-        break;
-    case balance_type::COPY_PRIMARY:
-        ans = "copy_primary";
-        result.balance_type = balancer_request_type::copy_primary;
-        result.action_list.emplace_back(
-            new_proposal_action(from, to, config_type::CT_ADD_SECONDARY_FOR_LB));
-        result.action_list.emplace_back(
-            new_proposal_action(from, from, config_type::CT_DOWNGRADE_TO_SECONDARY));
-        result.action_list.emplace_back(
-            new_proposal_action(to, to, config_type::CT_UPGRADE_TO_PRIMARY));
-        result.action_list.emplace_back(new_proposal_action(to, from, config_type::CT_REMOVE));
-        break;
-    case balance_type::COPY_SECONDARY:
-        ans = "copy_secondary";
-        result.balance_type = balancer_request_type::copy_secondary;
-        result.action_list.emplace_back(
-            new_proposal_action(pc.primary, to, config_type::CT_ADD_SECONDARY_FOR_LB));
-        result.action_list.emplace_back(
-            new_proposal_action(pc.primary, from, config_type::CT_REMOVE));
-        break;
-    default:
-        dassert(false, "");
-    }
-    ddebug("generate balancer: %d.%d %s from %s of disk_tag(%s) to %s",
-           pc.pid.get_app_id(),
-           pc.pid.get_partition_index(),
-           ans.c_str(),
-           from.to_string(),
-           get_disk_tag(apps, from, pc.pid).c_str(),
-           to.to_string());
-    return std::make_shared<configuration_balancer_request>(std::move(result));
-}
-
-void dump_disk_load(app_id id, const rpc_address &node, bool only_primary, const disk_load &load)
-{
-    std::ostringstream load_string;
-    load_string << std::endl << "<<<<<<<<<<" << std::endl;
-    load_string << "load for " << node.to_string() << ", "
-                << "app id: " << id;
-    if (only_primary) {
-        load_string << ", only for primary";
-    }
-    load_string << std::endl;
-
-    for (const auto &kv : load) {
-        load_string << kv.first << ": " << kv.second << std::endl;
-    }
-    load_string << ">>>>>>>>>>";
-    dinfo("%s", load_string.str().c_str());
-}
-
-bool calc_disk_load(node_mapper &nodes,
-                    const app_mapper &apps,
-                    app_id id,
-                    const rpc_address &node,
-                    bool only_primary,
-                    /*out*/ disk_load &load)
-{
-    load.clear();
-    const node_state *ns = get_node_state(nodes, node, false);
-    dassert(ns != nullptr, "can't find node(%s) from node_state", node.to_string());
-
-    auto add_one_replica_to_disk_load = [&](const gpid &pid) {
-        dinfo("add gpid(%d.%d) to node(%s) disk load",
-              pid.get_app_id(),
-              pid.get_partition_index(),
-              node.to_string());
-        const config_context &cc = *get_config_context(apps, pid);
-        auto iter = cc.find_from_serving(node);
-        if (iter == cc.serving.end()) {
-            dwarn("can't collect gpid(%d.%d)'s info from %s, which should be primary",
-                  pid.get_app_id(),
-                  pid.get_partition_index(),
-                  node.to_string());
-            return false;
-        } else {
-            load[iter->disk_tag]++;
-            return true;
-        }
-    };
-
-    if (only_primary) {
-        bool result = ns->for_each_primary(id, add_one_replica_to_disk_load);
-        dump_disk_load(id, node, only_primary, load);
-        return result;
-    } else {
-        bool result = ns->for_each_partition(id, add_one_replica_to_disk_load);
-        dump_disk_load(id, node, only_primary, load);
-        return result;
-    }
-}
-
 greedy_load_balancer::greedy_load_balancer(meta_service *_svc)
     : server_load_balancer(_svc),
       _ctrl_balancer_ignored_apps(nullptr),
@@ -244,6 +120,7 @@ greedy_load_balancer::greedy_load_balancer(meta_service *_svc)
         _only_primary_balancer = false;
         _only_move_primary = false;
     }
+    _app_balance_policy = dsn::make_unique<app_balance_policy>();
 
     ::memset(t_operation_counters, 0, sizeof(t_operation_counters));
 
@@ -269,64 +146,22 @@ greedy_load_balancer::greedy_load_balancer(meta_service *_svc)
         "copy secondary count by balancer in the recent period");
 }
 
-greedy_load_balancer::~greedy_load_balancer()
-{
-    UNREGISTER_VALID_HANDLER(_ctrl_balancer_in_turn);
-    UNREGISTER_VALID_HANDLER(_ctrl_only_primary_balancer);
-    UNREGISTER_VALID_HANDLER(_ctrl_only_move_primary);
-    UNREGISTER_VALID_HANDLER(_get_balance_operation_count);
-    UNREGISTER_VALID_HANDLER(_ctrl_balancer_ignored_apps);
-}
+greedy_load_balancer::~greedy_load_balancer() { unregister_ctrl_commands(); }
 
 void greedy_load_balancer::register_ctrl_commands()
 {
-    _ctrl_balancer_in_turn = dsn::command_manager::instance().register_command(
-        {"meta.lb.balancer_in_turn"},
-        "meta.lb.balancer_in_turn <true|false>",
-        "control whether do app balancer in turn",
-        [this](const std::vector<std::string> &args) {
-            return remote_command_set_bool_flag(_balancer_in_turn, "lb.balancer_in_turn", args);
-        });
-
-    _ctrl_only_primary_balancer = dsn::command_manager::instance().register_command(
-        {"meta.lb.only_primary_balancer"},
-        "meta.lb.only_primary_balancer <true|false>",
-        "control whether do only primary balancer",
-        [this](const std::vector<std::string> &args) {
-            return remote_command_set_bool_flag(
-                _only_primary_balancer, "lb.only_primary_balancer", args);
-        });
-
-    _ctrl_only_move_primary = dsn::command_manager::instance().register_command(
-        {"meta.lb.only_move_primary"},
-        "meta.lb.only_move_primary <true|false>",
-        "control whether only move primary in balancer",
-        [this](const std::vector<std::string> &args) {
-            return remote_command_set_bool_flag(_only_move_primary, "lb.only_move_primary", args);
-        });
-
     _get_balance_operation_count = dsn::command_manager::instance().register_command(
         {"meta.lb.get_balance_operation_count"},
         "meta.lb.get_balance_operation_count [total | move_pri | copy_pri | copy_sec | detail]",
         "get balance operation count",
         [this](const std::vector<std::string> &args) { return get_balance_operation_count(args); });
-
-    _ctrl_balancer_ignored_apps = dsn::command_manager::instance().register_command(
-        {"meta.lb.ignored_app_list"},
-        "meta.lb.ignored_app_list <get|set|clear> [app_id1,app_id2..]",
-        "get, set and clear balancer ignored_app_list",
-        [this](const std::vector<std::string> &args) {
-            return remote_command_balancer_ignored_app_ids(args);
-        });
+    _app_balance_policy->register_ctrl_commands();
 }
 
 void greedy_load_balancer::unregister_ctrl_commands()
 {
-    UNREGISTER_VALID_HANDLER(_ctrl_balancer_in_turn);
-    UNREGISTER_VALID_HANDLER(_ctrl_only_primary_balancer);
-    UNREGISTER_VALID_HANDLER(_ctrl_only_move_primary);
     UNREGISTER_VALID_HANDLER(_get_balance_operation_count);
-    UNREGISTER_VALID_HANDLER(_ctrl_balancer_ignored_apps);
+    _app_balance_policy->unregister_ctrl_commands();
 }
 
 std::string greedy_load_balancer::get_balance_operation_count(const std::vector<std::string> &args)
@@ -406,17 +241,6 @@ bool greedy_load_balancer::copy_primary(const std::shared_ptr<app_state> &app,
     return operation->start(t_migration_result);
 }
 
-bool greedy_load_balancer::copy_secondary(const std::shared_ptr<app_state> &app, bool place_holder)
-{
-    node_mapper &nodes = *(t_global_view->nodes);
-    const app_mapper &apps = *t_global_view->apps;
-    int replicas_low = app->partition_count / t_alive_nodes;
-
-    std::unique_ptr<copy_replica_operation> operation = dsn::make_unique<copy_secondary_operation>(
-        app, apps, nodes, address_vec, address_id, replicas_low);
-    return operation->start(t_migration_result);
-}
-
 void greedy_load_balancer::number_nodes(const node_mapper &nodes)
 {
     int current_id = 1;
@@ -430,140 +254,6 @@ void greedy_load_balancer::number_nodes(const node_mapper &nodes)
         address_id[iter->first] = current_id;
         address_vec[current_id] = iter->first;
         ++current_id;
-    }
-}
-
-ford_fulkerson::ford_fulkerson(const std::shared_ptr<app_state> &app,
-                               const node_mapper &nodes,
-                               const std::unordered_map<dsn::rpc_address, int> &address_id,
-                               uint32_t higher_count,
-                               uint32_t lower_count,
-                               int replicas_low)
-    : _app(app),
-      _nodes(nodes),
-      _address_id(address_id),
-      _higher_count(higher_count),
-      _lower_count(lower_count),
-      _replicas_low(replicas_low)
-{
-    make_graph();
-}
-
-// using dijstra to find shortest path
-std::unique_ptr<flow_path> ford_fulkerson::find_shortest_path()
-{
-    std::vector<bool> visit(_graph_nodes, false);
-    std::vector<int> flow(_graph_nodes, 0);
-    std::vector<int> prev(_graph_nodes, -1);
-    flow[0] = INT_MAX;
-    while (!visit.back()) {
-        auto pos = select_node(visit, flow);
-        if (pos == -1) {
-            break;
-        }
-        update_flow(pos, visit, _network, flow, prev);
-    }
-
-    if (visit.back() && flow.back() != 0) {
-        return dsn::make_unique<struct flow_path>(_app, std::move(flow), std::move(prev));
-    } else {
-        return nullptr;
-    }
-}
-
-void ford_fulkerson::make_graph()
-{
-    _graph_nodes = _nodes.size() + 2;
-    _network.resize(_graph_nodes, std::vector<int>(_graph_nodes, 0));
-    for (const auto &node : _nodes) {
-        int node_id = _address_id.at(node.first);
-        add_edge(node_id, node.second);
-        update_decree(node_id, node.second);
-    }
-    handle_corner_case();
-}
-
-void ford_fulkerson::add_edge(int node_id, const node_state &ns)
-{
-    int primary_count = ns.primary_count(_app->app_id);
-    if (primary_count > _replicas_low) {
-        _network[0][node_id] = primary_count - _replicas_low;
-    } else {
-        _network[node_id].back() = _replicas_low - primary_count;
-    }
-}
-
-void ford_fulkerson::update_decree(int node_id, const node_state &ns)
-{
-    ns.for_each_primary(_app->app_id, [&, this](const gpid &pid) {
-        const partition_configuration &pc = _app->partitions[pid.get_partition_index()];
-        for (const auto &secondary : pc.secondaries) {
-            auto i = _address_id.find(secondary);
-            dassert_f(i != _address_id.end(),
-                      "invalid secondary address, address = {}",
-                      secondary.to_string());
-            _network[node_id][i->second]++;
-        }
-        return true;
-    });
-}
-
-void ford_fulkerson::handle_corner_case()
-{
-    // Suppose you have an 8-shard app in a cluster with 3 nodes(which name are node1, node2,
-    // node3). The distribution of primaries among these nodes is as follow:
-    // node1 : [0, 1, 2, 3]
-    // node2 : [4, 5]
-    // node2 : [6, 7]
-    // This is obviously unbalanced.
-    // But if we don't handle this corner case, primary migration will not be triggered
-    if (_higher_count > 0 && _lower_count == 0) {
-        for (int i = 0; i != _graph_nodes; ++i) {
-            if (_network[0][i] > 0)
-                --_network[0][i];
-            else
-                ++_network[i][_graph_nodes - 1];
-        }
-    }
-}
-
-int ford_fulkerson::select_node(std::vector<bool> &visit, const std::vector<int> &flow)
-{
-    auto pos = max_value_pos(visit, flow);
-    if (pos != -1) {
-        visit[pos] = true;
-    }
-    return pos;
-}
-
-int ford_fulkerson::max_value_pos(const std::vector<bool> &visit, const std::vector<int> &flow)
-{
-    int pos = -1, max_value = 0;
-    for (auto i = 0; i != _graph_nodes; ++i) {
-        if (!visit[i] && flow[i] > max_value) {
-            pos = i;
-            max_value = flow[i];
-        }
-    }
-    return pos;
-}
-
-void ford_fulkerson::update_flow(int pos,
-                                 const std::vector<bool> &visit,
-                                 const std::vector<std::vector<int>> &network,
-                                 std::vector<int> &flow,
-                                 std::vector<int> &prev)
-{
-    for (auto i = 0; i != _graph_nodes; ++i) {
-        if (visit[i]) {
-            continue;
-        }
-
-        auto min = std::min(flow[pos], network[pos][i]);
-        if (min > flow[i]) {
-            flow[i] = min;
-            prev[i] = pos;
-        }
     }
 }
 
@@ -737,58 +427,13 @@ void greedy_load_balancer::greedy_balancer(const bool balance_checker)
     }
 
     if (!FLAGS_balance_cluster) {
-        app_balancer(balance_checker);
+        _app_balance_policy->balance(balance_checker, t_global_view, t_migration_result);
         return;
     }
 
     if (!balance_checker) {
         balance_cluster();
     }
-}
-
-void greedy_load_balancer::app_balancer(bool balance_checker)
-{
-    const app_mapper &apps = *t_global_view->apps;
-    if (!execute_balance(apps,
-                         balance_checker,
-                         _balancer_in_turn,
-                         _only_move_primary,
-                         std::bind(&greedy_load_balancer::primary_balance,
-                                   this,
-                                   std::placeholders::_1,
-                                   std::placeholders::_2))) {
-        return;
-    }
-
-    if (!need_balance_secondaries(balance_checker)) {
-        return;
-    }
-
-    // we seperate the primary/secondary balancer for 2 reasons:
-    // 1. globally primary balancer may make secondary unbalanced
-    // 2. in one-by-one mode, a secondary balance decision for an app may be prior than
-    // another app's primary balancer if not seperated.
-    execute_balance(apps,
-                    balance_checker,
-                    _balancer_in_turn,
-                    _only_move_primary,
-                    std::bind(&greedy_load_balancer::copy_secondary,
-                              this,
-                              std::placeholders::_1,
-                              std::placeholders::_2));
-}
-
-bool greedy_load_balancer::need_balance_secondaries(bool balance_checker)
-{
-    if (!balance_checker && !t_migration_result->empty()) {
-        ddebug("stop to do secondary balance coz we already has actions to do");
-        return false;
-    }
-    if (_only_primary_balancer) {
-        ddebug("stop to do secondary balancer coz it is not allowed");
-        return false;
-    }
-    return true;
 }
 
 bool greedy_load_balancer::execute_balance(
@@ -1407,147 +1052,6 @@ bool greedy_load_balancer::is_ignored_app(app_id app_id)
     return _balancer_ignored_apps.find(app_id) != _balancer_ignored_apps.end();
 }
 
-std::unordered_map<dsn::rpc_address, disk_load>
-get_node_loads(const std::shared_ptr<app_state> &app,
-               const app_mapper &apps,
-               node_mapper &nodes,
-               bool only_primary)
-{
-    std::unordered_map<dsn::rpc_address, disk_load> node_loads;
-    for (auto iter = nodes.begin(); iter != nodes.end(); ++iter) {
-        if (!calc_disk_load(
-                nodes, apps, app->app_id, iter->first, only_primary, node_loads[iter->first])) {
-            dwarn_f("stop the balancer as some replica infos aren't collected, node({}), app({})",
-                    iter->first.to_string(),
-                    app->get_logname());
-            return node_loads;
-        }
-    }
-    return node_loads;
-}
-
-copy_replica_operation::copy_replica_operation(
-    const std::shared_ptr<app_state> app,
-    const app_mapper &apps,
-    node_mapper &nodes,
-    const std::vector<dsn::rpc_address> &address_vec,
-    const std::unordered_map<dsn::rpc_address, int> &address_id)
-    : _app(app), _apps(apps), _nodes(nodes), _address_vec(address_vec), _address_id(address_id)
-{
-}
-
-bool copy_replica_operation::start(migration_list *result)
-{
-    init_ordered_address_ids();
-    _node_loads = get_node_loads(_app, _apps, _nodes, only_copy_primary());
-    if (_node_loads.size() != _nodes.size()) {
-        return false;
-    }
-
-    while (true) {
-        if (!can_continue()) {
-            break;
-        }
-
-        gpid selected_pid = select_partition(result);
-        if (selected_pid.get_app_id() != -1) {
-            copy_once(selected_pid, result);
-            update_ordered_address_ids();
-        } else {
-            _ordered_address_ids.erase(--_ordered_address_ids.end());
-        }
-    }
-    return true;
-}
-
-const partition_set *copy_replica_operation::get_all_partitions()
-{
-    int id_max = *_ordered_address_ids.rbegin();
-    const node_state &ns = _nodes.find(_address_vec[id_max])->second;
-    const partition_set *partitions = ns.partitions(_app->app_id, only_copy_primary());
-    return partitions;
-}
-
-gpid copy_replica_operation::select_max_load_gpid(const partition_set *partitions,
-                                                  migration_list *result)
-{
-    int id_max = *_ordered_address_ids.rbegin();
-    const disk_load &load_on_max = _node_loads.at(_address_vec[id_max]);
-
-    gpid selected_pid(-1, -1);
-    int max_load = -1;
-    for (const gpid &pid : *partitions) {
-        if (!can_select(pid, result)) {
-            continue;
-        }
-
-        const std::string &disk_tag = get_disk_tag(_apps, _address_vec[id_max], pid);
-        auto load = load_on_max.at(disk_tag);
-        if (load > max_load) {
-            selected_pid = pid;
-            max_load = load;
-        }
-    }
-    return selected_pid;
-}
-
-void copy_replica_operation::copy_once(gpid selected_pid, migration_list *result)
-{
-    auto from = _address_vec[*_ordered_address_ids.rbegin()];
-    auto to = _address_vec[*_ordered_address_ids.begin()];
-
-    auto pc = _app->partitions[selected_pid.get_partition_index()];
-    auto request = generate_balancer_request(_apps, pc, get_balance_type(), from, to);
-    result->emplace(selected_pid, request);
-}
-
-void copy_replica_operation::update_ordered_address_ids()
-{
-    int id_min = *_ordered_address_ids.begin();
-    int id_max = *_ordered_address_ids.rbegin();
-    --_partition_counts[id_max];
-    ++_partition_counts[id_min];
-
-    _ordered_address_ids.erase(_ordered_address_ids.begin());
-    _ordered_address_ids.erase(--_ordered_address_ids.end());
-
-    _ordered_address_ids.insert(id_max);
-    _ordered_address_ids.insert(id_min);
-}
-
-void copy_replica_operation::init_ordered_address_ids()
-{
-    _partition_counts.resize(_address_vec.size(), 0);
-    for (const auto &iter : _nodes) {
-        auto id = _address_id.at(iter.first);
-        _partition_counts[id] = get_partition_count(iter.second);
-    }
-
-    std::set<int, std::function<bool(int a, int b)>> ordered_queue([this](int a, int b) {
-        return _partition_counts[a] != _partition_counts[b]
-                   ? _partition_counts[a] < _partition_counts[b]
-                   : a < b;
-    });
-    for (const auto &iter : _nodes) {
-        auto id = _address_id.at(iter.first);
-        ordered_queue.insert(id);
-    }
-    _ordered_address_ids.swap(ordered_queue);
-}
-
-gpid copy_replica_operation::select_partition(migration_list *result)
-{
-    const partition_set *partitions = get_all_partitions();
-
-    int id_max = *_ordered_address_ids.rbegin();
-    const node_state &ns = _nodes.find(_address_vec[id_max])->second;
-    dassert_f(partitions != nullptr && !partitions->empty(),
-              "max load({}) shouldn't empty",
-              ns.addr().to_string());
-
-    return select_max_load_gpid(partitions, result);
-}
-
 copy_primary_operation::copy_primary_operation(
     const std::shared_ptr<app_state> app,
     const app_mapper &apps,
@@ -1591,71 +1095,5 @@ bool copy_primary_operation::can_continue()
 }
 
 enum balance_type copy_primary_operation::get_balance_type() { return balance_type::COPY_PRIMARY; }
-
-copy_secondary_operation::copy_secondary_operation(
-    const std::shared_ptr<app_state> app,
-    const app_mapper &apps,
-    node_mapper &nodes,
-    const std::vector<dsn::rpc_address> &address_vec,
-    const std::unordered_map<dsn::rpc_address, int> &address_id,
-    int replicas_low)
-    : copy_replica_operation(app, apps, nodes, address_vec, address_id), _replicas_low(replicas_low)
-{
-}
-
-bool copy_secondary_operation::can_continue()
-{
-    int id_min = *_ordered_address_ids.begin();
-    int id_max = *_ordered_address_ids.rbegin();
-    if (_partition_counts[id_max] <= _replicas_low ||
-        _partition_counts[id_max] - _partition_counts[id_min] <= 1) {
-        ddebug_f("{}: stop copy secondary coz it will be balanced later", _app->get_logname());
-        return false;
-    }
-    return true;
-}
-
-int copy_secondary_operation::get_partition_count(const node_state &ns) const
-{
-    return ns.partition_count(_app->app_id);
-}
-
-bool copy_secondary_operation::can_select(gpid pid, migration_list *result)
-{
-    int id_max = *_ordered_address_ids.rbegin();
-    const node_state &max_ns = _nodes.at(_address_vec[id_max]);
-    if (max_ns.served_as(pid) == partition_status::PS_PRIMARY) {
-        dinfo_f("{}: skip gpid({}.{}) coz it is primary",
-                _app->get_logname(),
-                pid.get_app_id(),
-                pid.get_partition_index());
-        return false;
-    }
-
-    // if the pid have been used
-    if (result->find(pid) != result->end()) {
-        dinfo_f("{}: skip gpid({}.{}) coz it is already copyed",
-                _app->get_logname(),
-                pid.get_app_id(),
-                pid.get_partition_index());
-        return false;
-    }
-
-    int id_min = *_ordered_address_ids.begin();
-    const node_state &min_ns = _nodes.at(_address_vec[id_min]);
-    if (min_ns.served_as(pid) != partition_status::PS_INACTIVE) {
-        dinfo_f("{}: skip gpid({}.{}) coz it is already a member on the target node",
-                _app->get_logname(),
-                pid.get_app_id(),
-                pid.get_partition_index());
-        return false;
-    }
-    return true;
-}
-
-enum balance_type copy_secondary_operation::get_balance_type()
-{
-    return balance_type::COPY_SECONDARY;
-}
 } // namespace replication
 } // namespace dsn
