+++ src/meta/test/copy_replica_operation_test.cpp
@@ -17,6 +17,7 @@
 
 #include <gtest/gtest.h>
 #include <dsn/utility/fail_point.h>
+#include "meta/app_balance_policy.h"
 #include "meta/greedy_load_balancer.h"
 
 namespace dsn {
